Add log formatter to file appender
